Move related classes to distinct package
