+++ src/main/java/org/jabref/gui/frame/OpenConsoleAction.java
@@ -1,9 +1,11 @@
-package org.jabref.gui;
+package org.jabref.gui.frame;
 
 import java.io.IOException;
 import java.util.Optional;
 import java.util.function.Supplier;
 
+import org.jabref.gui.DialogService;
+import org.jabref.gui.StateManager;
 import org.jabref.gui.actions.ActionHelper;
 import org.jabref.gui.actions.SimpleCommand;
 import org.jabref.gui.desktop.JabRefDesktop;
