+++ src/test/java/org/jabref/gui/util/OpenConsoleActionTest.java
@@ -2,8 +2,8 @@ package org.jabref.gui.util;
 
 import java.util.Optional;
 
-import org.jabref.gui.OpenConsoleAction;
 import org.jabref.gui.StateManager;
+import org.jabref.gui.frame.OpenConsoleAction;
 import org.jabref.model.database.BibDatabaseContext;
 import org.jabref.preferences.PreferencesService;
